Bug 7720: create/delete VPN CLI handling addition/removal of subnets
Following changes have been done as part of the bug fix:

1. Addition of subnets to VPN done via associateNetworks RPC
2. Deletion of subnets from VPN done via dissociateNetworks RPC
3. (1) is done only if createVPN succeeds
4. Deletion of VPN is done only if (2) succeeds

+++ vpnservice/neutronvpn/neutronvpn-shell/src/main/java/org/opendaylight/netvirt/neutronvpn/shell/ConfigureL3VpnCommand.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016 Ericsson India Global Services Pvt Ltd. and others.  All rights reserved.
+ * Copyright (c) 2016, 2017 Ericsson India Global Services Pvt Ltd. and others.  All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
@@ -14,10 +14,14 @@ import org.apache.karaf.shell.console.OsgiCommandSupport;
 import org.opendaylight.controller.sal.binding.api.RpcProviderRegistry;
 import org.opendaylight.netvirt.neutronvpn.interfaces.INeutronVpnManager;
 import org.opendaylight.yang.gen.v1.urn.ietf.params.xml.ns.yang.ietf.yang.types.rev130715.Uuid;
+import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.neutronvpn.rev150602.AssociateNetworksInputBuilder;
+import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.neutronvpn.rev150602.AssociateNetworksOutput;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.neutronvpn.rev150602.CreateL3VPNInputBuilder;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.neutronvpn.rev150602.CreateL3VPNOutput;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.neutronvpn.rev150602.DeleteL3VPNInputBuilder;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.neutronvpn.rev150602.DeleteL3VPNOutput;
+import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.neutronvpn.rev150602.DissociateNetworksInputBuilder;
+import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.neutronvpn.rev150602.DissociateNetworksOutput;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.neutronvpn.rev150602.NeutronvpnService;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.neutronvpn.rev150602.createl3vpn.input.L3vpn;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.netvirt.neutronvpn.rev150602.createl3vpn.input.L3vpnBuilder;
@@ -137,6 +141,7 @@ public class ConfigureL3VpnCommand extends OsgiCommandSupport {
 
         Uuid vuuid = new Uuid(vid);
 
+        RpcResult<CreateL3VPNOutput> createL3VpnRpcResult = null;
         try {
             ArrayList<String> rdList = new ArrayList<>(Arrays.asList(rd.split(",")));
             ArrayList<String> irtList = new ArrayList<>(Arrays.asList(irt.split(",")));
@@ -148,18 +153,18 @@ public class ConfigureL3VpnCommand extends OsgiCommandSupport {
             }
 
             List<L3vpn> l3vpns = new ArrayList<>();
-            L3vpn l3vpn = new L3vpnBuilder().setId(vuuid).setName(name).setRouteDistinguisher(rdList).setImportRT
-                    (irtList)
-                    .setExportRT(ertList).setTenantId(tuuid).build();
+
+            L3vpn l3vpn = new L3vpnBuilder().setId(vuuid).setName(name).setRouteDistinguisher(rdList)
+                    .setImportRT(irtList).setExportRT(ertList).setTenantId(tuuid).build();
             l3vpns.add(l3vpn);
             Future<RpcResult<CreateL3VPNOutput>> result =
                     neutronvpnService.createL3VPN(new CreateL3VPNInputBuilder().setL3vpn(l3vpns).build());
-            RpcResult<CreateL3VPNOutput> rpcResult = result.get();
-            if (rpcResult.isSuccessful()) {
+            createL3VpnRpcResult = result.get();
+            if (createL3VpnRpcResult.isSuccessful()) {
                 session.getConsole().println("L3VPN created successfully");
                 Logger.trace("createl3vpn: {}", result);
             } else {
-                session.getConsole().println("error populating createL3VPN : " + result.get().getErrors());
+                session.getConsole().println("Error populating createL3VPN : " + result.get().getErrors());
                 session.getConsole().println(getHelp("create"));
             }
         } catch (InterruptedException | ExecutionException e) {
@@ -168,22 +173,46 @@ public class ConfigureL3VpnCommand extends OsgiCommandSupport {
             session.getConsole().println(getHelp("create"));
         }
 
-        try {
-            List<Uuid> sidList = new ArrayList<>();
+        /**
+         * passing a subnetId list alongwith create-l3-vpn CLI implicitly indicates that
+         * association of network(s) to VPN is being intended.
+         */
+        if (createL3VpnRpcResult.isSuccessful()) {
+            try {
+                List<Uuid> networkIdList = new ArrayList<>();
 
-            if (sid != null) {
-                for (String sidStr : sid.split(",")) {
-                    Uuid suuid = new Uuid(sidStr);
-                    sidList.add(suuid);
+                if (sid != null) {
+                    for (String sidStr : sid.split(",")) {
+                        Uuid subnetId = new Uuid(sidStr);
+                        Uuid networkId = neutronVpnManager.getNetworkForSubnet(subnetId);
+                        if (networkId != null) {
+                            networkIdList.add(networkId);
+                        } else {
+                            session.getConsole().println("Could not find network for subnet " + subnetId.getValue()
+                                    + ". Not proceeding with adding subnet to VPN");
+                        }
+                    }
+
+                    if (!networkIdList.isEmpty()) {
+                        Future<RpcResult<AssociateNetworksOutput>> result =
+                                neutronvpnService.associateNetworks(new AssociateNetworksInputBuilder()
+                                        .setVpnId(vuuid).setNetworkId(networkIdList).build());
+                        RpcResult<AssociateNetworksOutput> associateNetworksRpcResult = result.get();
+                        if (associateNetworksRpcResult.isSuccessful()) {
+                            session.getConsole().println("Subnet(s) added to VPN successfully");
+                            Logger.trace("associateNetworks: {}", result);
+                        } else {
+                            session.getConsole().println("Error while adding subnet(s) to VPN: "
+                                    + result.get().getErrors());
+                            session.getConsole().println(getHelp("create"));
+                        }
+                    }
                 }
+            } catch (InterruptedException | ExecutionException e) {
+                Logger.error("error while adding subnet(s) to VPN", e);
+                session.getConsole().println("Error while adding subnet(s) to VPN: " + e.getMessage());
+                session.getConsole().println(getHelp("create"));
             }
-            for (Uuid subnet : sidList) {
-                neutronVpnManager.addSubnetToVpn(vuuid, subnet);
-            }
-
-        } catch (Exception e) {
-            Logger.error("error in adding subnet to VPN", e);
-            session.getConsole().println("error in adding subnet to VPN : " + e.getMessage());
         }
     }
 
@@ -194,37 +223,57 @@ public class ConfigureL3VpnCommand extends OsgiCommandSupport {
             session.getConsole().println(getHelp("delete"));
             return;
         }
-        Uuid vpnid = new Uuid(vid);
+        Uuid vpnId = new Uuid(vid);
+
+        // disassociation of network(s) (removal of subnet(s)) from VPN to be followed by deletion of VPN
+        RpcResult<DissociateNetworksOutput> dissociateNetworksRpcResult = null;
+        List<Uuid> networkIdList = null;
         try {
-            List<Uuid> sidList = neutronVpnManager.getSubnetsforVpn(vpnid);
-            if (sidList != null) {
-                for (Uuid subnet : sidList) {
-                    neutronVpnManager.removeSubnetFromVpn(vpnid, subnet);
+            networkIdList = neutronVpnManager.getNetworksForVpn(vpnId);
+
+            if (networkIdList != null && !networkIdList.isEmpty()) {
+                Future<RpcResult<DissociateNetworksOutput>> result =
+                        neutronvpnService.dissociateNetworks(new DissociateNetworksInputBuilder()
+                                .setVpnId(vpnId).setNetworkId(networkIdList).build());
+                dissociateNetworksRpcResult = result.get();
+                if (dissociateNetworksRpcResult.isSuccessful()) {
+                    session.getConsole().println("Subnet(s) removed from VPN successfully");
+                    Logger.trace("dissociateNetworks: {}", result);
+                } else {
+                    session.getConsole().println("Error while removing subnet(s) from VPN: "
+                            + result.get().getErrors());
+                    session.getConsole().println(getHelp("delete"));
                 }
             }
-        } catch (Exception e) {
-            Logger.error("error in deleting subnet from VPN", e);
-            session.getConsole().println("error in deleting subnet from VPN : " + e.getMessage());
+        } catch (InterruptedException | ExecutionException e) {
+            Logger.error("error while adding removing subnet(s) from VPN", e);
+            session.getConsole().println("Error while removing subnet(s) from VPN: " + e.getMessage());
+            session.getConsole().println(getHelp("delete"));
         }
 
-        try {
-            List<Uuid> vpnIdList = new ArrayList<>();
-            vpnIdList.add(vpnid);
+        if (networkIdList == null || networkIdList.isEmpty() || dissociateNetworksRpcResult.isSuccessful()) {
+            try {
+                List<Uuid> vpnIdList = new ArrayList<>();
+                vpnIdList.add(vpnId);
 
-            Future<RpcResult<DeleteL3VPNOutput>> result =
-                    neutronvpnService.deleteL3VPN(new DeleteL3VPNInputBuilder().setId(vpnIdList).build());
-            RpcResult<DeleteL3VPNOutput> rpcResult = result.get();
-            if (rpcResult.isSuccessful()) {
-                session.getConsole().println("L3VPN deleted successfully");
-                Logger.trace("deletel3vpn: {}", result);
-            } else {
-                session.getConsole().println("error populating deleteL3VPN : " + result.get().getErrors());
+                Future<RpcResult<DeleteL3VPNOutput>> result =
+                        neutronvpnService.deleteL3VPN(new DeleteL3VPNInputBuilder().setId(vpnIdList).build());
+                RpcResult<DeleteL3VPNOutput> rpcResult = result.get();
+                if (rpcResult.isSuccessful()) {
+                    session.getConsole().println("L3VPN deleted successfully");
+                    Logger.trace("deletel3vpn: {}", result);
+                } else {
+                    session.getConsole().println("Error populating deleteL3VPN : " + result.get().getErrors());
+                    session.getConsole().println(getHelp("delete"));
+                }
+            } catch (InterruptedException | ExecutionException e) {
+                Logger.error("error populating deleteL3VPN", e);
+                session.getConsole().println("Error populating deleteL3VPN : " + e.getMessage());
                 session.getConsole().println(getHelp("delete"));
             }
-        } catch (InterruptedException | ExecutionException e) {
-            Logger.error("error populating deleteL3VPN", e);
-            session.getConsole().println("error populating deleteL3VPN : " + e.getMessage());
-            session.getConsole().println(getHelp("delete"));
+        } else {
+            session.getConsole().println("Not proceeding with deletion of L3VPN since error(s) encountered "
+                    + "in removing subnet(s) from VPN");
         }
     }
